refactor(lotto): Lotto 파일 이동
